Split Postgres schema alignment tests
Move shared Postgres schema-alignment harness and catalogue assertions into a
helper module, and isolate article-threading coverage behind focused phase
helpers.

Rename the SQLite `add_sn` fresh-insert test so its name and setup comment
match the zero-default behaviour it asserts.
