--- src/db/schema_alignment_tests/postgres_tests.rs
@@ -1,481 +1,0 @@
-//! `PostgreSQL` schema alignment regression tests.
-
-use std::future::Future;
-
-use diesel::sql_query;
-use diesel_async::{AsyncConnection, RunQueryDsl};
-use test_util::postgres::PostgresTestDb;
-
-use super::{
-    DbConnection,
-    NameRow,
-    TestResult,
-    apply_migrations,
-    assert_permission_round_trip_with_ids,
-    assert_root_category_names_are_unique,
-    assert_upgrade_backfills,
-    setup_legacy_schema,
-};
-
-async fn setup_postgres_legacy_schema(conn: &mut DbConnection) -> TestResult<()> {
-    setup_legacy_schema(
-        conn,
-        &[
-            include_str!("../../../migrations/postgres/00000000000000_create_users/up.sql"),
-            include_str!("../../../migrations/postgres/00000000000001_create_news/up.sql"),
-            include_str!("../../../migrations/postgres/00000000000002_add_bundles/up.sql"),
-            include_str!("../../../migrations/postgres/00000000000003_add_articles/up.sql"),
-            include_str!("../../../migrations/postgres/00000000000004_create_files/up.sql"),
-            include_str!(
-                "../../../migrations/postgres/00000000000005_add_bundle_name_parent_index/up.sql"
-            ),
-        ],
-    )
-    .await
-}
-
-async fn postgres_names(conn: &mut DbConnection, sql: &str) -> TestResult<Vec<String>> {
-    Ok(sql_query(sql)
-        .load::<NameRow>(conn)
-        .await?
-        .into_iter()
-        .map(|row| row.name)
-        .collect())
-}
-
-async fn assert_postgres_permission_schema(conn: &mut DbConnection) -> TestResult<()> {
-    let permission_tables = postgres_names(
-        conn,
-        "SELECT table_name AS name FROM information_schema.tables WHERE table_schema = 'public' \
-         AND table_name IN ('permissions', 'user_permissions') ORDER BY table_name",
-    )
-    .await?;
-    assert_eq!(permission_tables, vec!["permissions", "user_permissions"]);
-
-    let permission_indices = postgres_names(
-        conn,
-        "SELECT indexname AS name FROM pg_indexes WHERE schemaname = 'public' AND tablename IN \
-         ('permissions', 'user_permissions') ORDER BY indexname",
-    )
-    .await?;
-    for expected in ["permissions_code_key", "user_permissions_pkey"] {
-        assert!(permission_indices.iter().any(|name| name == expected));
-    }
-    Ok(())
-}
-
-async fn assert_postgres_bundle_schema(conn: &mut DbConnection) -> TestResult<()> {
-    let bundle_columns = postgres_names(
-        conn,
-        "SELECT column_name AS name FROM information_schema.columns WHERE table_name = \
-         'news_bundles' AND table_schema = 'public' ORDER BY ordinal_position",
-    )
-    .await?;
-    assert_eq!(
-        bundle_columns,
-        vec!["id", "parent_bundle_id", "name", "guid", "created_at"]
-    );
-
-    let bundle_indices = postgres_names(
-        conn,
-        "SELECT indexname AS name FROM pg_indexes WHERE tablename = 'news_bundles' AND schemaname \
-         = 'public' ORDER BY indexname",
-    )
-    .await?;
-    for expected in [
-        "idx_bundles_name_parent",
-        "idx_bundles_parent",
-        "news_bundles_name_parent_bundle_id_key",
-    ] {
-        assert!(bundle_indices.iter().any(|name| name == expected));
-    }
-
-    let bundle_constraints = postgres_names(
-        conn,
-        "SELECT conname AS name FROM pg_constraint WHERE conrelid = \
-         'public.news_bundles'::regclass AND contype = 'u' ORDER BY conname",
-    )
-    .await?;
-    assert!(
-        bundle_constraints
-            .iter()
-            .any(|name| name == "news_bundles_name_parent_bundle_id_key")
-    );
-    Ok(())
-}
-
-async fn assert_postgres_category_schema(conn: &mut DbConnection) -> TestResult<()> {
-    let category_columns = postgres_names(
-        conn,
-        "SELECT column_name AS name FROM information_schema.columns WHERE table_name = \
-         'news_categories' AND table_schema = 'public' ORDER BY ordinal_position",
-    )
-    .await?;
-    assert_eq!(
-        category_columns,
-        vec![
-            "id",
-            "name",
-            "bundle_id",
-            "guid",
-            "add_sn",
-            "delete_sn",
-            "created_at"
-        ]
-    );
-
-    let category_indices = postgres_names(
-        conn,
-        "SELECT indexname AS name FROM pg_indexes WHERE tablename = 'news_categories' AND \
-         schemaname = 'public' ORDER BY indexname",
-    )
-    .await?;
-    for expected in [
-        "idx_categories_bundle",
-        "idx_categories_root_name_unique",
-        "idx_categories_name_bundle_unique",
-    ] {
-        assert!(category_indices.iter().any(|name| name == expected));
-    }
-    Ok(())
-}
-
-/// Verifies that `news_articles` has the expected `PostgreSQL` indexes.
-///
-/// The check is order-agnostic: it queries index names with `postgres_names`
-/// through the supplied `DbConnection` and asserts that `idx_articles_category`,
-/// `idx_articles_first_child_article`, `idx_articles_next_article`,
-/// `idx_articles_parent_article`, and `idx_articles_prev_article` are present.
-/// Database query failures are returned as `TestResult<()>`; missing indexes
-/// panic through the assertions.
-pub(super) async fn assert_postgres_article_indices(conn: &mut DbConnection) -> TestResult<()> {
-    let article_indices = postgres_names(
-        conn,
-        "SELECT indexname AS name FROM pg_indexes WHERE tablename = 'news_articles' AND \
-         schemaname = 'public' ORDER BY indexname",
-    )
-    .await?;
-    for expected in [
-        "idx_articles_category",
-        "idx_articles_first_child_article",
-        "idx_articles_next_article",
-        "idx_articles_parent_article",
-        "idx_articles_prev_article",
-    ] {
-        assert!(article_indices.iter().any(|name| name == expected));
-    }
-    Ok(())
-}
-
-async fn assert_postgres_news_schema(conn: &mut DbConnection) -> TestResult<()> {
-    assert_postgres_bundle_schema(conn).await?;
-    assert_postgres_category_schema(conn).await?;
-    assert_postgres_article_indices(conn).await
-}
-
-async fn assert_postgres_aligned_schema(conn: &mut DbConnection) -> TestResult<()> {
-    assert_postgres_permission_schema(conn).await?;
-    assert_postgres_news_schema(conn).await?;
-    assert_root_category_names_are_unique(conn).await
-}
-
-async fn assert_permission_round_trip(conn: &mut DbConnection) -> TestResult<()> {
-    assert_permission_round_trip_with_ids(
-        conn,
-        super::PermissionTestIds {
-            user_id: 42,
-            permission_id: 42,
-            code: 34,
-        },
-    )
-    .await
-}
-
-fn is_ci() -> bool { std::env::var("CI").is_ok_and(|value| !value.is_empty()) }
-
-fn with_postgres_test_db<F, Fut>(test: F) -> TestResult<()>
-where
-    F: FnOnce(String) -> Fut + Send + 'static,
-    Fut: Future<Output = TestResult<()>> + Send + 'static,
-{
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()?;
-    runtime.block_on(async move {
-        if std::env::var_os("POSTGRES_TEST_URL").is_some() {
-            let db = PostgresTestDb::new_async().await?;
-            return test(db.url.to_string()).await;
-        }
-
-        run_with_embedded_postgres(test).await
-    })
-}
-
-async fn run_with_embedded_postgres<F, Fut>(test: F) -> TestResult<()>
-where
-    F: FnOnce(String) -> Fut + Send + 'static,
-    Fut: Future<Output = TestResult<()>> + Send + 'static,
-{
-    let Some(pg) = start_optional_embedded_postgres().await? else {
-        return Ok(());
-    };
-
-    let db_name = format!(
-        "schema_alignment_{}_{}",
-        std::process::id(),
-        std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)?
-            .as_nanos()
-    );
-    let result = async {
-        pg.create_database(&db_name).await?;
-        let url = pg.settings().url(&db_name);
-        test(url).await
-    }
-    .await;
-
-    let stop_result = stop_embedded_postgres(pg);
-    combine_postgres_test_result(result, stop_result)
-}
-
-async fn start_optional_embedded_postgres() -> TestResult<Option<postgresql_embedded::PostgreSQL>> {
-    let mut pg = postgresql_embedded::PostgreSQL::default();
-    if let Err(error) = pg.setup().await {
-        if is_ci() {
-            return Err(error.into());
-        }
-        tracing::warn!("SKIP-TEST-CLUSTER: PostgreSQL unavailable");
-        return Ok(None);
-    }
-    pg.start().await?;
-    Ok(Some(pg))
-}
-
-fn stop_embedded_postgres(pg: postgresql_embedded::PostgreSQL) -> TestResult<()> {
-    std::thread::spawn(move || {
-        let runtime = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()?;
-        runtime.block_on(pg.stop()).map_err(anyhow::Error::from)
-    })
-    .join()
-    .map_err(|_| anyhow::anyhow!("embedded postgres shutdown thread panicked"))?
-}
-
-fn combine_postgres_test_result(
-    result: TestResult<()>,
-    stop_result: TestResult<()>,
-) -> TestResult<()> {
-    match (result, stop_result) {
-        (Ok(()), Ok(())) => Ok(()),
-        (Err(test_error), Ok(())) => Err(test_error),
-        (Ok(()), Err(stop_error)) => Err(stop_error),
-        (Err(test_error), Err(stop_error)) => Err(anyhow::anyhow!(
-            "postgres schema alignment test failed: {test_error}; embedded postgres shutdown \
-             failed: {stop_error}"
-        )),
-    }
-}
-
-#[serial_test::file_serial(postgres_embedded_setup)]
-#[test]
-fn postgres_fresh_migration_creates_aligned_schema() -> TestResult<()> {
-    with_postgres_test_db(|url| async move {
-        let mut conn = DbConnection::establish(&url).await?;
-        apply_migrations(&mut conn, &url, None).await?;
-
-        assert_postgres_aligned_schema(&mut conn).await?;
-        assert_permission_round_trip(&mut conn).await
-    })
-}
-
-#[serial_test::file_serial(postgres_embedded_setup)]
-#[test]
-fn postgres_upgrade_backfills_legacy_news_rows() -> TestResult<()> {
-    with_postgres_test_db(|url| async move {
-        let mut conn = DbConnection::establish(&url).await?;
-        setup_postgres_legacy_schema(&mut conn).await?;
-        apply_migrations(&mut conn, &url, None).await?;
-
-        assert_upgrade_backfills(&mut conn).await?;
-        assert_postgres_aligned_schema(&mut conn).await
-    })
-}
-
-#[expect(clippy::panic_in_result_fn, reason = "test assertions")]
-#[serial_test::file_serial(postgres_embedded_setup)]
-#[test]
-fn postgres_category_names_are_bundle_scoped() -> TestResult<()> {
-    with_postgres_test_db(|url| async move {
-        let mut conn = DbConnection::establish(&url).await?;
-        apply_migrations(&mut conn, &url, None).await?;
-
-        sql_query("INSERT INTO news_bundles (parent_bundle_id, name) VALUES (NULL, 'BundleA')")
-            .execute(&mut conn)
-            .await?;
-        sql_query("INSERT INTO news_bundles (parent_bundle_id, name) VALUES (NULL, 'BundleB')")
-            .execute(&mut conn)
-            .await?;
-
-        let bundle_ids = postgres_names(
-            &mut conn,
-            "SELECT id::text AS name FROM news_bundles ORDER BY id",
-        )
-        .await?;
-        let (bid1, bid2) = (&bundle_ids[0], &bundle_ids[1]);
-
-        // Same name in different bundles must succeed
-        sql_query(format!(
-            "INSERT INTO news_categories (name, bundle_id) VALUES ('Sports', {bid1})"
-        ))
-        .execute(&mut conn)
-        .await?;
-        sql_query(format!(
-            "INSERT INTO news_categories (name, bundle_id) VALUES ('Sports', {bid2})"
-        ))
-        .execute(&mut conn)
-        .await?;
-
-        // Same name in the same bundle must fail
-        let duplicate = sql_query(format!(
-            "INSERT INTO news_categories (name, bundle_id) VALUES ('Sports', {bid1})"
-        ))
-        .execute(&mut conn)
-        .await;
-        assert!(
-            duplicate.is_err(),
-            "duplicate name in same bundle must be rejected"
-        );
-        Ok(())
-    })
-}
-
-#[expect(clippy::panic_in_result_fn, reason = "test assertions")]
-#[serial_test::file_serial(postgres_embedded_setup)]
-#[test]
-fn postgres_guids_are_non_empty_and_unique() -> TestResult<()> {
-    with_postgres_test_db(|url| async move {
-        let mut conn = DbConnection::establish(&url).await?;
-        apply_migrations(&mut conn, &url, None).await?;
-
-        sql_query(
-            "INSERT INTO news_bundles (parent_bundle_id, name) VALUES (NULL, 'GA'), (NULL, 'GB')",
-        )
-        .execute(&mut conn)
-        .await?;
-
-        let guids = postgres_names(
-            &mut conn,
-            "SELECT guid AS name FROM news_bundles ORDER BY id",
-        )
-        .await?;
-        assert_eq!(guids.len(), 2, "expected two bundle rows");
-        for guid in &guids {
-            assert!(!guid.is_empty(), "GUID must not be empty");
-        }
-        assert_ne!(guids[0], guids[1], "GUIDs must be unique across rows");
-        Ok(())
-    })
-}
-
-#[expect(clippy::panic_in_result_fn, reason = "test assertions")]
-#[serial_test::file_serial(postgres_embedded_setup)]
-#[test]
-fn postgres_article_threading_enforces_referential_integrity() -> TestResult<()> {
-    with_postgres_test_db(|url| async move {
-        let mut conn = DbConnection::establish(&url).await?;
-        apply_migrations(&mut conn, &url, None).await?;
-
-        sql_query(
-            "INSERT INTO news_bundles (parent_bundle_id, name) VALUES (NULL, 'ThreadBundle')",
-        )
-        .execute(&mut conn)
-        .await?;
-
-        let bundle_ids = postgres_names(
-            &mut conn,
-            "SELECT id::text AS name FROM news_bundles ORDER BY id",
-        )
-        .await?;
-        let bid = &bundle_ids[0];
-
-        sql_query(format!(
-            "INSERT INTO news_categories (name, bundle_id) VALUES ('ThreadCat', {bid})"
-        ))
-        .execute(&mut conn)
-        .await?;
-
-        let cat_ids = postgres_names(
-            &mut conn,
-            "SELECT id::text AS name FROM news_categories ORDER BY id",
-        )
-        .await?;
-        let cid = &cat_ids[0];
-
-        // Root article
-        sql_query(format!(
-            "INSERT INTO news_articles (category_id, parent_article_id, prev_article_id, \
-             next_article_id, first_child_article_id, title, posted_at) VALUES ({cid}, NULL, \
-             NULL, NULL, NULL, 'Root', NOW())"
-        ))
-        .execute(&mut conn)
-        .await?;
-
-        let root_ids = postgres_names(
-            &mut conn,
-            "SELECT id::text AS name FROM news_articles ORDER BY id",
-        )
-        .await?;
-        let rid = &root_ids[0];
-
-        // Child article referencing root via parent_article_id
-        sql_query(format!(
-            "INSERT INTO news_articles (category_id, parent_article_id, prev_article_id, \
-             next_article_id, first_child_article_id, title, posted_at) VALUES ({cid}, {rid}, \
-             NULL, NULL, NULL, 'Child', NOW())"
-        ))
-        .execute(&mut conn)
-        .await?;
-
-        let child_ids = postgres_names(
-            &mut conn,
-            "SELECT id::text AS name FROM news_articles WHERE parent_article_id IS NOT NULL ORDER \
-             BY id",
-        )
-        .await?;
-        assert_eq!(child_ids.len(), 1, "expected one child article");
-        let chid = &child_ids[0];
-
-        // Update root to point first_child_article_id at child
-        sql_query(format!(
-            "UPDATE news_articles SET first_child_article_id = {chid} WHERE id = {rid}"
-        ))
-        .execute(&mut conn)
-        .await?;
-
-        // Verify the threading link via a JOIN query
-        let linked = postgres_names(
-            &mut conn,
-            &format!(
-                "SELECT a.id::text AS name FROM news_articles a INNER JOIN news_articles child ON \
-                 child.id = a.first_child_article_id WHERE a.id = {rid}"
-            ),
-        )
-        .await?;
-        assert_eq!(linked.len(), 1, "root article must link to its child");
-        assert_eq!(linked[0], *rid, "linked root id must match");
-
-        // Referential integrity: non-existent parent_article_id must be rejected
-        let bad_insert = sql_query(format!(
-            "INSERT INTO news_articles (category_id, parent_article_id, title, posted_at) VALUES \
-             ({cid}, 999999, 'Orphan', NOW())"
-        ))
-        .execute(&mut conn)
-        .await;
-        assert!(
-            bad_insert.is_err(),
-            "insert with non-existent parent_article_id must be rejected"
-        );
-        Ok(())
-    })
-}
